+++ tests/Common/CommonQueryBuilderTest.php
@@ -7,7 +7,7 @@ namespace Yiisoft\Db\Tests\Common;
 use PHPUnit\Framework\Attributes\DoesNotPerformAssertions;
 use Yiisoft\Db\Command\CommandInterface;
 use Yiisoft\Db\Exception\Exception;
-use Yiisoft\Db\Schema\Column\ColumnSchemaInterface;
+use Yiisoft\Db\Schema\Column\ColumnInterface;
 use Yiisoft\Db\Tests\AbstractQueryBuilderTest;
 use Yiisoft\Db\Tests\Provider\QueryBuilderProvider;
 
@@ -34,7 +34,7 @@ abstract class CommonQueryBuilderTest extends AbstractQueryBuilderTest
         foreach ($provider as $data) {
             $column = $data[1];
 
-            if ($column instanceof ColumnSchemaInterface) {
+            if ($column instanceof ColumnInterface) {
                 if ($column->isPrimaryKey()) {
                     $this->createTebleWithColumn($command, $column);
                     continue;
@@ -59,7 +59,7 @@ abstract class CommonQueryBuilderTest extends AbstractQueryBuilderTest
         $command->createTable('build_column_definition', $columns)->execute();
     }
 
-    private function createTebleWithColumn(CommandInterface $command, string|ColumnSchemaInterface $column)
+    private function createTebleWithColumn(CommandInterface $command, string|ColumnInterface $column)
     {
         try {
             $command->dropTable('build_column_definition_primary_key')->execute();