+++ CHANGELOG.md
@@ -52,6 +52,7 @@
 - New #906: Add `ServerInfoInterface` and its implementation (@Tigrov)
 - Enh #905: Use `AbstractColumnDefinitionBuilder` to generate table column SQL representation (@Tigrov)
 - Enh #915: Remove `ColumnInterface` (@Tigrov)
+- Enh #917: Rename `ColumnSchemaInterface` to `ColumnInterface` (@Tigrov)
 
 ## 1.3.0 March 21, 2024